refactor(patina): split no_arrow_functions_in_watch tests into submodule
